Use spreadsheet instead of `fs` for event image urls
Downsides are that this doesn't update immediately when adding an image locally (and doesn't show images for events that aren't done being written up yet), but this lets event images be ordered by start date instead of the semi-random order they were in before. Plus, it saves a call to `readdirSync`.